feat: add auth route and protected routes

+++ routes/item.py
@@ -1,13 +1,18 @@
 from flask import Blueprint, jsonify,request
 from models.db import db
 from models.item import Item
+from flask_jwt_extended import jwt_required
+from routes.utils.items import is_authorized
 
 items_bp = Blueprint("items",__name__,url_prefix="/items")
 
 @items_bp.route("/<int:item_id>", methods=["PUT"])
+@jwt_required()
 def update_order(item_id):
     try:
         item = db.get_or_404(Item, item_id)
+        if not is_authorized(item.order_id):
+            return jsonify({"msg": "No autorizado"}), 403
         data = request.get_json()
         garment_variant_id = data.get("garment_variant_id")
         design_id = data.get("design_id")
@@ -23,9 +28,13 @@ def update_order(item_id):
 
 
 @items_bp.route("/<int:item_id>", methods=["DELETE"])
+@jwt_required()
 def delete_item( item_id):
     try:
         item= db.get_or_404(Item, item_id)  
+        if not is_authorized(item.order_id):
+            return jsonify({"msg": "No autorizado"}), 403
+
         db.session.delete(item) 
         db.session.commit() 
         return jsonify({"message": "Item deleted"}), 200
